수정 요구사항 반영
