+++ week04/ToDoList/ToDoLIst/TaskList.csv
@@ -3,3 +3,6 @@
 밥먹기,COMPLETION
 양치하기,COMPLETION
 TIL쓰기,PROCESSING
+청소하기,COMPLETION
+안경닦기,DELETION
+화장실 청소하기,PROCESSING
