+++ week04/ToDoList/ToDoLIst/src/main/java/TaskLoader.java
@@ -2,6 +2,8 @@ import models.Task;
 
 import java.io.File;
 import java.io.FileNotFoundException;
+import java.io.FileWriter;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
@@ -23,6 +25,16 @@ public class TaskLoader {
     }
 
     return tasks;
+  }
+
+  public void writeTask(List<Task> tasks) throws IOException {
+    FileWriter fileWriter = new FileWriter("TaskList.csv");
+    for (Task task : tasks) {
+      String line = task.toCsvRow();
+
+      fileWriter.write(line + "\n");
+    }
+    fileWriter.close();
   }
 
   private Task parseTask(String text) {
