+++ week04/ToDoList/ToDoLIst/src/main/java/panels/TasksPanel.java
@@ -21,41 +21,43 @@ public class TasksPanel extends JPanel {
 
   public void addContentPanel() {
     for (Task task : tasks) {
-      if (!task.state().equals("DELETION")) {
-        JPanel taskPanel = new JPanel();
+      if(task.state().equals("DELETION")) {
+        continue;
+      }
 
-        JCheckBox taskCheckBox = new JCheckBox();
-        if (task.state().equals("COMPLETION")) {
+      JPanel taskPanel = new JPanel();
+
+      JCheckBox taskCheckBox = new JCheckBox();
+      if (task.state().equals("COMPLETION")) {
+        taskCheckBox.setSelected(true);
+      }
+
+      taskCheckBox.addActionListener(event -> {
+        task.complete();
+      });
+      taskPanel.add(taskCheckBox);
+
+      JLabel taskTitleLabel = new JLabel(task.toString());
+      taskTitleLabel.addMouseListener(new MouseAdapter() {
+        public void mouseClicked(MouseEvent e) {
+          if (taskCheckBox.isSelected()) {
+            taskCheckBox.setSelected(false);
+            return;
+          }
           taskCheckBox.setSelected(true);
         }
+      });
+      taskPanel.add(taskTitleLabel);
 
-        taskCheckBox.addActionListener(event -> {
-          task.complete();
-        });
-        taskPanel.add(taskCheckBox);
+      JButton taskDeleteButton = new JButton("X");
+      taskDeleteButton.addActionListener(event -> {
+        task.delete();
 
-        JLabel taskTitleLabel = new JLabel(task.toString());
-        taskTitleLabel.addMouseListener(new MouseAdapter() {
-          public void mouseClicked(MouseEvent e) {
-            if (taskCheckBox.isSelected()) {
-              taskCheckBox.setSelected(false);
-              return;
-            }
-            taskCheckBox.setSelected(true);
-          }
-        });
-        taskPanel.add(taskTitleLabel);
+        refreshTaskPanel();
+      });
+      taskPanel.add(taskDeleteButton);
 
-        JButton taskDeleteButton = new JButton("X");
-        taskDeleteButton.addActionListener(event -> {
-          task.delete();
-
-          refreshTaskPanel();
-        });
-        taskPanel.add(taskDeleteButton);
-
-        this.add(taskPanel);
-      }
+      this.add(taskPanel);
     }
   }
